Move benchmarks to main test files
